+++ src/Sentry.Unity.Editor/SentryCli.cs
@@ -106,7 +106,7 @@ namespace Sentry.Unity.Editor
         internal static string? UrlOverride(string? dsnOption, string? urlOverrideOption)
         {
             string? result = urlOverrideOption;
-            if (result is null && !string.IsNullOrEmpty(dsnOption))
+            if (string.IsNullOrEmpty(result) && !string.IsNullOrEmpty(dsnOption))
             {
                 var uri = new Uri(dsnOption);
 
@@ -116,7 +116,7 @@ namespace Sentry.Unity.Editor
                     result = new UriBuilder(uri.Scheme, uri.DnsSafeHost, uri.Port, "").Uri.AbsoluteUri.TrimEnd('/');
                 }
             }
-            return result;
+            return string.IsNullOrEmpty(result) ? null : result;
         }
     }
 }